Made file names consistent, add main function for bot, modified much of memers_markov to follow style guidelines (mostly docstrings)
